model_training/requirements.txt
# Model Training Dependencies for Cattle Breed Identification

# Core ML Libraries
tensorflow>=2.13.0
tensorflow-hub>=0.14.0
tensorflowjs>=4.10.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Kaggle API
kaggle>=1.5.16

# Image Augmentation
albumentations>=1.3.0
imgaug>=0.4.0

# Progress Tracking
tqdm>=4.65.0

# Utilities
pathlib2>=2.3.7
requests>=2.31.0
